+++ sekoia_automation/aio/connector.py
@@ -1,7 +1,7 @@
 """Contains connector with async version."""
 
+import asyncio
 from abc import ABC
-from asyncio import AbstractEventLoop, get_event_loop
 from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
 from datetime import datetime
@@ -21,8 +21,6 @@ class AsyncConnector(Connector, ABC):
 
     configuration: DefaultConnectorConfiguration
 
-    _event_loop: AbstractEventLoop
-
     _session: ClientSession | None = None
     _rate_limiter: AsyncLimiter | None = None
 
@@ -30,7 +28,6 @@ class AsyncConnector(Connector, ABC):
         self,
         module: Module | None = None,
         data_path: Path | None = None,
-        event_loop: AbstractEventLoop | None = None,
         *args,
         **kwargs,
     ):
@@ -47,52 +44,49 @@ class AsyncConnector(Connector, ABC):
         self.max_concurrency_tasks = kwargs.pop("max_concurrency_tasks", 1000)
         super().__init__(module=module, data_path=data_path, *args, **kwargs)
 
-        self._event_loop = event_loop or get_event_loop()
-
-    @classmethod
-    def set_client_session(cls, session: ClientSession) -> None:
+    def set_client_session(self, session: ClientSession) -> None:
         """
         Set client session.
 
         Args:
             session: ClientSession
         """
-        cls._session = session
+        self._session = session
 
-    @classmethod
-    def set_rate_limiter(cls, rate_limiter: AsyncLimiter) -> None:
+    def set_rate_limiter(self, rate_limiter: AsyncLimiter) -> None:
         """
         Set rate limiter.
 
         Args:
             rate_limiter:
         """
-        cls._rate_limiter = rate_limiter
+        self._rate_limiter = rate_limiter
 
-    @classmethod
-    def get_rate_limiter(cls) -> AsyncLimiter:
+    def get_rate_limiter(self) -> AsyncLimiter:
         """
         Get or initialize rate limiter.
 
         Returns:
             AsyncLimiter:
         """
-        if cls._rate_limiter is None:
-            cls._rate_limiter = AsyncLimiter(1, 1)
+        if self._rate_limiter is None:
+            self._rate_limiter = AsyncLimiter(1, 1)
 
-        return cls._rate_limiter
+        return self._rate_limiter
 
-    @classmethod
     @asynccontextmanager
-    async def session(cls) -> AsyncGenerator[ClientSession, None]:  # pragma: no cover
+    async def session(self) -> AsyncGenerator[ClientSession, None]:  # pragma: no cover
         """
         Get or initialize client session if it is not initialized yet.
 
         Returns:
             ClientSession:
         """
-        async with ClientSession() as cls._session, cls.get_rate_limiter():
-            yield cls._session
+        if self._session is None:
+            self._session = ClientSession()
+
+        async with self.get_rate_limiter():
+            yield self._session
 
     async def _async_send_chunk(
         self, session: ClientSession, url: str, chunk_index: int, chunk: list[str]
@@ -166,3 +160,14 @@ class AsyncConnector(Connector, ABC):
                 result_ids.extend(ids)
 
         return result_ids
+
+    def stop(self, *args, **kwargs):
+        """
+        Stop the connector
+        """
+        loop = asyncio.get_event_loop()
+
+        if self._session:
+            loop.run_until_complete(self._session.close())
+
+        super().stop(*args, **kwargs)