+++ tests/aio/test_connector.py
@@ -52,15 +52,12 @@ async def test_async_connector_rate_limiter(async_connector: DummyAsyncConnector
     assert async_connector._rate_limiter is None
     assert other_instance._rate_limiter is None
 
-    assert async_connector.get_rate_limiter() == other_instance.get_rate_limiter()
+    assert async_connector.get_rate_limiter() != other_instance.get_rate_limiter()
 
     async_connector.set_rate_limiter(rate_limiter_mock)
 
-    assert async_connector.get_rate_limiter() == other_instance.get_rate_limiter()
+    assert async_connector.get_rate_limiter() != other_instance.get_rate_limiter()
     assert async_connector._rate_limiter == rate_limiter_mock
-
-    DummyAsyncConnector.set_rate_limiter(None)
-    DummyAsyncConnector.set_client_session(None)
 
 
 @pytest.mark.asyncio
@@ -76,17 +73,17 @@ async def test_async_connector_client_session(async_connector: DummyAsyncConnect
     assert async_connector._session is None
     assert other_instance._session is None
 
-    async with async_connector.session():
-        assert async_connector._rate_limiter is not None and isinstance(
-            async_connector._rate_limiter, AsyncLimiter
-        )
+    async with async_connector.session() as session_1:
+        async with other_instance.session() as session_2:
+            assert session_1 != session_2
 
-        assert other_instance._rate_limiter is not None and isinstance(
-            other_instance._rate_limiter, AsyncLimiter
-        )
+            assert async_connector._rate_limiter is not None and isinstance(
+                async_connector._rate_limiter, AsyncLimiter
+            )
 
-    DummyAsyncConnector.set_rate_limiter(None)
-    other_instance.set_client_session(None)
+            assert other_instance._rate_limiter is not None and isinstance(
+                other_instance._rate_limiter, AsyncLimiter
+            )
 
 
 @pytest.mark.asyncio